moved server functions into separate files
