+++ server/getCredentials.ts
@@ -2,7 +2,7 @@
 
 import { Credentials } from "@/app/types";
 
-const validLoginDetails: Credentials[] = [
+export const validLoginDetails: Credentials[] = [
   {
     username: "FakeUser1",
     password: "Password1",
@@ -13,20 +13,3 @@ const validLoginDetails: Credentials[] = [
   },
 ];
 
-export async function credentialsAreValid(
-  inputUsername: string,
-  inputPassword: string
-) {
-  return validLoginDetails.some((loginDetails, index) => {
-    return (
-      loginDetails.username === inputUsername &&
-      validLoginDetails[index].password === inputPassword
-    );
-  });
-}
-
-export async function userNameAlreadyExists(inputUsername: string) {
-  return validLoginDetails.some((loginDetails) => {
-    return loginDetails.username === inputUsername;
-  });
-}
