+++ server/usernameAlreadyExists.ts
@@ -1,0 +1,9 @@
+"use server";
+
+import { validLoginDetails } from "./getCredentials";
+
+export async function usernameAlreadyExists(inputUsername: string) {
+  return validLoginDetails.some((loginDetails) => {
+    return loginDetails.username === inputUsername;
+  });
+}
